Add http status code in response.

+++ aiccs-20191015/src/main/java/com/aliyun/aiccs20191015/models/GetHotlineAgentDetailReportRequest.java
@@ -20,10 +20,10 @@ public class GetHotlineAgentDetailReportRequest extends TeaModel {
     public String instanceId;
 
     @NameInMap("DepIds")
-    public java.util.List<Integer> depIds;
+    public java.util.List<Long> depIds;
 
     @NameInMap("GroupIds")
-    public java.util.List<Integer> groupIds;
+    public java.util.List<Long> groupIds;
 
     public static GetHotlineAgentDetailReportRequest build(java.util.Map<String, ?> map) throws Exception {
         GetHotlineAgentDetailReportRequest self = new GetHotlineAgentDetailReportRequest();
@@ -70,19 +70,19 @@ public class GetHotlineAgentDetailReportRequest extends TeaModel {
         return this.instanceId;
     }
 
-    public GetHotlineAgentDetailReportRequest setDepIds(java.util.List<Integer> depIds) {
+    public GetHotlineAgentDetailReportRequest setDepIds(java.util.List<Long> depIds) {
         this.depIds = depIds;
         return this;
     }
-    public java.util.List<Integer> getDepIds() {
+    public java.util.List<Long> getDepIds() {
         return this.depIds;
     }
 
-    public GetHotlineAgentDetailReportRequest setGroupIds(java.util.List<Integer> groupIds) {
+    public GetHotlineAgentDetailReportRequest setGroupIds(java.util.List<Long> groupIds) {
         this.groupIds = groupIds;
         return this;
     }
-    public java.util.List<Integer> getGroupIds() {
+    public java.util.List<Long> getGroupIds() {
         return this.groupIds;
     }
